+++ .circleci/config.yml
@@ -23,15 +23,15 @@ aliases:
     command: |
       set -exu
 
-      mkdir -p packages/react-recommendations/dist
+      mkdir -p packages/js-horizontal-slider/dist
       mkdir -p packages/js-recommendations/dist
-      mkdir -p packages/horizontal-slider-js/dist
-      mkdir -p packages/horizontal-slider-react/dist
+      mkdir -p packages/react-horizontal-slider/dist
+      mkdir -p packages/react-recommendations/dist
 
-      cp -R /tmp/workspace/packages/react-recommendations/dist packages/react-recommendations
+      cp -R /tmp/workspace/packages/js-horizontal-slider/dist packages/js-horizontal-slider
       cp -R /tmp/workspace/packages/js-recommendations/dist packages/js-recommendations
-      cp -R /tmp/workspace/packages/horizontal-slider-js/dist packages/horizontal-slider-js
-      cp -R /tmp/workspace/packages/horizontal-slider-react/dist packages/horizontal-slider-react
+      cp -R /tmp/workspace/packages/react-horizontal-slider/dist packages/react-horizontal-slider
+      cp -R /tmp/workspace/packages/react-recommendations/dist packages/react-recommendations
 
 defaults: &defaults
   working_directory: ~/ui-components
@@ -61,15 +61,15 @@ jobs:
           command: |
             set -exu
 
-            mkdir -p /tmp/workspace/packages/react-recommendations/dist
+            mkdir -p /tmp/workspace/packages/js-horizontal-slider/dist
             mkdir -p /tmp/workspace/packages/js-recommendations/dist
-            mkdir -p /tmp/workspace/packages/horizontal-slider-js/dist
-            mkdir -p /tmp/workspace/packages/horizontal-slider-react/dist
+            mkdir -p /tmp/workspace/packages/react-horizontal-slider/dist
+            mkdir -p /tmp/workspace/packages/react-recommendations/dist
 
-            cp -R packages/react-recommendations/dist /tmp/workspace/packages/react-recommendations
+            cp -R packages/js-horizontal-slider/dist /tmp/workspace/packages/js-horizontal-slider
             cp -R packages/js-recommendations/dist /tmp/workspace/packages/js-recommendations
-            cp -R packages/horizontal-slider-js/dist /tmp/workspace/packages/horizontal-slider-js
-            cp -R packages/horizontal-slider-react/dist /tmp/workspace/packages/horizontal-slider-react
+            cp -R packages/react-horizontal-slider/dist /tmp/workspace/packages/react-horizontal-slider
+            cp -R packages/react-recommendations/dist /tmp/workspace/packages/react-recommendations
       - persist_to_workspace:
           root: *workspace_root
           paths:
