+++ examples/demo/package.json
@@ -12,7 +12,7 @@
     "@algolia/autocomplete-js": "1.0.0",
     "@algolia/autocomplete-theme-classic": "1.0.0",
     "@algolia/react-recommendations": "1.0.0-experimental.0",
-    "@algolia/ui-components-horizontal-slider-react": "1.0.0-experimental.0",
+    "@algolia/ui-components-react-horizontal-slider": "1.0.0-experimental.0",
     "algoliasearch": "4.9.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
