+++ examples/js-demo/package.json
@@ -15,7 +15,7 @@
     "@algolia/autocomplete-js": "1.0.1",
     "@algolia/autocomplete-theme-classic": "1.0.1",
     "@algolia/js-recommendations": "1.0.0-experimental.0",
-    "@algolia/ui-components-horizontal-slider-js": "1.0.0-experimental.0",
+    "@algolia/ui-components-js-horizontal-slider": "1.0.0-experimental.0",
     "algoliasearch": "4.9.1",
     "preact": "10.5.13",
     "search-insights": "1.8.0"
