+++ packages/js-horizontal-slider/rollup.config.js
