+++ packages/js-horizontal-slider/src/getHTMLElement.ts
