+++ packages/js-horizontal-slider/src/index.ts
