+++ packages/js-horizontal-slider/tsconfig.declaration.json
