+++ packages/react-horizontal-slider/HorizontalSlider.css
