+++ packages/react-horizontal-slider/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@algolia/ui-components-horizontal-slider-react",
+  "name": "@algolia/ui-components-react-horizontal-slider",
   "description": "Horizontal slider UI component.",
   "version": "1.0.0-experimental.0",
   "license": "MIT",
