+++ packages/react-horizontal-slider/rollup.config.js
