+++ packages/react-horizontal-slider/src/HorizontalSlider.tsx
