+++ packages/react-horizontal-slider/src/index.ts
