+++ packages/react-horizontal-slider/tsconfig.declaration.json
