+++ ship.config.js
@@ -9,8 +9,8 @@ module.exports = {
     packagesToPublish: [
       'packages/react-recommendations',
       'packages/js-recommendations',
-      'packages/horizontal-slider-react',
-      'packages/horizontal-slider-js',
+      'packages/react-horizontal-slider',
+      'packages/js-horizontal-slider',
     ],
   },
   publishCommand({ tag }) {
